+++ test/unit/fuzz-tools.test.js
@@ -319,6 +319,7 @@ import * as cormackLehaneV339 from '../../lib/cormack-lehane-v339.js';
 import * as clarkLevelV340 from '../../lib/clark-level-v340.js';
 import * as masonRadialHeadV341 from '../../lib/mason-radial-head-v341.js';
 import * as hawkinsTalarV342 from '../../lib/hawkins-talar-v342.js';
+import * as sandersCalcanealV343 from '../../lib/sanders-calcaneal-v343.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -583,6 +584,7 @@ const MODULES = {
   'clark-level-v340.js': clarkLevelV340,
   'mason-radial-head-v341.js': masonRadialHeadV341,
   'hawkins-talar-v342.js': hawkinsTalarV342,
+  'sanders-calcaneal-v343.js': sandersCalcanealV343,
   'tb-testing.js': tbTesting,
 };